backend/requirements.txt
Flask>=2.0
pymongo>=4.0
python-dateutil>=2.8
azure-functions>=1.11
gunicorn
